+++ R/ovlp_sels.R
@@ -12,12 +12,12 @@
 #' @param pb Logical argument to control progress bar and messages. Default is \code{TRUE}.
 #' @param max.ovlp Numeric vector of length 1 specifying the maximum overlap allowed (in seconds)
 #' . Default is 0. 
-#' @param relabel Logical. If \code{TRUE} then selections are names (selec column) are reset.  
+#' @param relabel Logical. If \code{TRUE} then selections names (selec column) are reset.  
 #' Default is \code{FALSE}.
 #' @param drop Logical. If \code{TRUE}, when 2 or more selections overlap the function will remove 
 #' all but one of the overlapping selection. Default is \code{FALSE}.
 #' @param priority Character vector. Controls the priority criteria used for removing overlapped selections. It
-#' must list the levels of the column used to determine priority (argument priority.col) in the desire 
+#' must list the levels of the column used to determine priority (argument priority.col) in the desired 
 #' priority order. Default is \code{NULL}.
 #' @param priority.col Character vector of length with the name of the column use to determine the priority of
 #' overlapped selections. Default is \code{NULL}.
@@ -34,7 +34,7 @@
 #'
 #' # modified selec.table to make the first and second selection overlap
 #' Y <- selec.table
-#' Y$end[4] <- 1.5 
+#' Y$end[4] <- 1.5 S
 #'   
 #'  ovlp_sels(X =  Y)
 #' 
@@ -44,7 +44,8 @@
 #' # get index instead
 #'  ovlp_sels(X =  Y, index = TRUE)
 #' }
-#' @details This function detects selections overlapping in time. Selections must be liste in a data frame similar to \code{\link{selec.table}}.
+#' @details This function detects selections within a selection table that overlap in time. Selections must be 
+#' listed in a data frame similar to \code{\link{selec.table}}.
 #' @seealso  \code{\link{filtersels}} \code{\link{selec.table}}
 #' @author Marcelo Araya-Salas (\email{araya-salas@@cornell.edu})
 #last modification on may-22-2017 (MAS)